folder and file restructure
