+++ package.json
@@ -10,8 +10,11 @@
     "moment": "^2.29.4",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
-    "react-router-dom": "^6.18.0",
+    "react-icons": "^4.12.0",
+    "react-moment": "^1.1.3",
+    "react-router-dom": "^6.20.1",
     "react-scripts": "5.0.1",
+    "sass": "^1.69.5",
     "typed.js": "^2.1.0",
     "web-vitals": "^2.1.4"
   },
